Demo 6 complete
I finished Demo 6. The issue with loading the texture was that I needed it in a server.  Instead of creating a server (like a normal developer), I stole a texture already online, and referenced that instead.  On a side note, it looks like there maybe a way to load the image without a server, but I still can't find what it is.